fixed renaming folder bugs
