+++ src/components/666Workout/DisplayExercice.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import { Link } from 'react-router-dom'
 import ReactPlayer from 'react-player'
-import Directions from './Directions.js'
+import Directions from './Directions'
 import Loader from '../Loader'
 import '../../App.scss'
